[3D] Hide sideways faces on the 3D Frame.
A small follow-up for #6603. When a 3D frame face has `visible === 'auto'`, it should only be visible if it is facing the user.

But the behavior when the face is 90º from the user isn't very clear -- Should it be shown or not?
My initial decision was to show it, but I changed my mind -- It looks bad, doesn't display any information and sometimes gets on the way of axis labels.

This patch also adds some protection agains rounding errors